+++ src/test/Ejector.t.sol
@@ -8,13 +8,13 @@ contract EjectorTest is TestBase {
     constructor() {}
 
     function testEjector() public {
-        uint256 nftId = nft.mint();
+        uint nftId = nft.mint();
         currency.mint(address(bidderWithEjector), MIN_AMOUNT_FOR_NEW_CLONE);
 
         vm.startPrank(address(bidderWithEjector));
         currency.approve(dmAddr, MIN_AMOUNT_FOR_NEW_CLONE);
 
-        (uint256 cloneId, /*uint256 protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
+        (uint cloneId, /*uint protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
         vm.stopPrank();
 
         uint128 minAmountToBuyClone = dm.getMinAmountForCloneTransfer(cloneId);
@@ -25,19 +25,19 @@ contract EjectorTest is TestBase {
         dm.duplicate(nftAddr, nftId, currencyAddr, minAmountToBuyClone, false, 0);
         vm.stopPrank();
 
-        uint256 ejections = bidderWithEjector.ejections();
+        uint ejections = bidderWithEjector.ejections();
         assertEq(ejections, 1);
         assertEq(dm.ownerOf(cloneId), eoa1);
     }
 
     function testEjectorWithRevert() public {
-        uint256 nftId = nft.mint();
+        uint nftId = nft.mint();
         currency.mint(address(bidderWithBadEjector), MIN_AMOUNT_FOR_NEW_CLONE);
 
         vm.startPrank(address(bidderWithBadEjector));
         currency.approve(dmAddr, MIN_AMOUNT_FOR_NEW_CLONE);
 
-        (uint256 cloneId, /*uint256 protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
+        (uint cloneId, /*uint protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
         vm.stopPrank();
 
         uint128 minAmountToBuyClone = dm.getMinAmountForCloneTransfer(cloneId);
@@ -48,19 +48,19 @@ contract EjectorTest is TestBase {
         dm.duplicate(nftAddr, nftId, currencyAddr, minAmountToBuyClone, false, 0);
         vm.stopPrank();
 
-        uint256 ejections = bidderWithBadEjector.ejections();
+        uint ejections = bidderWithBadEjector.ejections();
         assertEq(ejections, 0);
         assertEq(dm.ownerOf(cloneId), eoa1);
     }
 
     function testEjectorWithOOG() public {
-        uint256 nftId = nft.mint();
+        uint nftId = nft.mint();
         currency.mint(address(bidderWithGassyEjector), MIN_AMOUNT_FOR_NEW_CLONE);
 
         vm.startPrank(address(bidderWithGassyEjector));
         currency.approve(dmAddr, MIN_AMOUNT_FOR_NEW_CLONE);
 
-        (uint256 cloneId, /*uint256 protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
+        (uint cloneId, /*uint protoId*/) = dm.duplicate(nftAddr, nftId, currencyAddr, MIN_AMOUNT_FOR_NEW_CLONE, false, 0);
         vm.stopPrank();
 
         uint128 minAmountToBuyClone = dm.getMinAmountForCloneTransfer(cloneId);
@@ -71,7 +71,7 @@ contract EjectorTest is TestBase {
         dm.duplicate(nftAddr, nftId, currencyAddr, minAmountToBuyClone, false, 0);
         vm.stopPrank();
 
-        uint256 ejections = bidderWithGassyEjector.ejections();
+        uint ejections = bidderWithGassyEjector.ejections();
         assertEq(ejections, 0);
         assertEq(dm.ownerOf(cloneId), eoa1);
     }
